% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{block2}
\alias{block2}
\title{Psychology-subset of National Institute of Mental Health's (NIMH) Volunteer Dataset}
\format{
A data frame with 183 rows and 5 variables:
\describe{
\item{Alcohol}{A numeric vector representing how often subject had a drink containing alcohol (0 - never; 1 - monthly or less; 2 - 2-4 times a week; 3 - 2-3 times a week; 4 - 4 or more times a week).}
\item{Verbal_IQ}{A numeric vector representing standardized verbal IQ score}
\item{Nonverbal_IQ}{A numeric vector representing standardized non-verbal IQ score}
\item{ME_fat}{A numeric vector representing mental fatigue score}
\item{ME_per}{A numeric vector representing mental performance}
}
}
\source{
\url{https://doi.org/doi:10.18112/openneuro.ds004215.v1.0.2}
}
\usage{
block2
}
\description{
A "psychology" subset of the National Institute of Mental Health's Healthy Volunteer data set (version 1.0.2) that contains AUDIT (alcohol use disorder), KBIT-2 (intelligence), and VAS (mental effort) measures.
All variables have been nonparanormal transformed.
}
\references{
Nugent, A. C., Thomas, A. G., Mahoney, M., Gibbons, A., Smith, J. T., Charles, A. J., ... & Chung, J. (2022). The NIMH intramural healthy volunteer dataset: A comprehensive MEG, MRI, and behavioral resource. \emph{Scientific Data}, 9(1), 518.
}
\keyword{datasets}
