% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootnet_SNAC.R
\name{bootnet_SNAC}
\alias{bootnet_SNAC}
\title{Bootstrapped SNAC common network estimation}
\usage{
bootnet_SNAC(
  data,
  nBoots = 1000,
  type = "nonparametric",
  nCores = 1,
  statistics = "edge",
  fun,
  verbose = TRUE,
  labels,
  alpha = 1,
  caseMin = 0.05,
  caseMax = 0.75,
  caseN = 10,
  subCases,
  computeCentrality = TRUE,
  propBoot = 1,
  replacement = TRUE,
  includeDiagonal = FALSE,
  bridgeArgs = list(),
  library = .libPaths(),
  memorysaver = TRUE,
  ...
)
}
\arguments{
\item{data}{Data set}

\item{nBoots}{Number of bootstraps}

\item{type}{Bootstrap method to use, default to "nonparametric"}

\item{nCores}{Number of cores used, default is 1}

\item{statistics}{Type of statistics calculated, defaults to "edge"}

\item{fun}{Function used to estimate network}

\item{verbose}{Messages on what is being done?}

\item{labels}{Variable labels}

\item{alpha}{centrality alpha}

\item{caseMin}{Minimum proportion to drop}

\item{caseMax}{Maximum proportion to drop}

\item{caseN}{Default is 10}

\item{subCases}{subCases}

\item{computeCentrality}{Whether to compute centrality TRUE/FALSE}

\item{propBoot}{M out of N}

\item{replacement}{Bootstrap with replacement? Default is TRUE}

\item{includeDiagonal}{Include diagonal? Default is FALSE}

\item{bridgeArgs}{Argument for bridge statistics}

\item{library}{Path to save files}

\item{memorysaver}{Default is TRUE}

\item{...}{Argument for 'fun' function}
}
\value{
A bootnet object with the following elements
\item{sampleTable}{A data frame containing all estimated values on the real sample.}
\item{bootTable}{A data frame containing all estimated values on the bootstrapped samples.}
\item{sample}{A bootnetResult object with plot and print method containing the estimated network of the real sample.}
\item{boots}{A list of bootnetResult objects containing the raw bootstrap results.}
\item{numbCommon}{Number of times the component comp was a common component.}
\item{lambdaCom}{String of bootstrapped 10-fold cross-validation values for lasso lambda.}
\item{lambdaNet}{String of bootstrapped 10-fold cross-validation values for graphical lasso lambda.}
}
\description{
Adjusted version of function bootnet() from R package bootnet (Epskamp, Borsboom, & Fried, 2018)
}
\references{
Epskamp, S., Borsboom, D., & Fried, E. I. (2018). Estimating psychological networks and their accuracy: A tutorial paper. Behavior Research Methods, 50(1), 195-212.
}
