% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootSNAC.R
\name{bootSNAC}
\alias{bootSNAC}
\title{Single function to perform two-step Sparse Network And Component (SNAC) analysis to estimate cross-domain partial correlations}
\usage{
bootSNAC(
  data,
  b1,
  b2,
  R,
  method = "structured",
  Target,
  comp,
  lambda_com,
  lambda_net
)
}
\arguments{
\item{data}{Data file consisting of 2 data blocks with subjects as rows and variables as columns. Variables should be grouped per data block}

\item{b1}{The number of variables in data block 1}

\item{b2}{The number of variables in data block 2}

\item{R}{The number of components}

\item{method}{How is component structure determined? Default is "structured" for which a predetermined component structure in the form of a Target is required.}

\item{Target}{Required for structured analysis. Should be matrix of 2 rows (for each data block) by R components. Enter 0 if data block should not contribute to that component; 1 is data block should contribute to that component.}

\item{comp}{The position of the component to bootstrap}

\item{lambda_com}{Lambda parameter for lasso (component step). Default is determined by 10-fold cross-validation.}

\item{lambda_net}{Lambda parameter for graphical lasso (network step). Default is determined by 10-fold cross-validation.}
}
\value{
\item{graph}{A partial correlation matrix}
\item{inv_cov}{Inverse covariance matrix.}
\item{cv_lambda_com}{Estimated value for lasso lambda.}
\item{cv_lambda_net}{Estimated value for graphical lasso lambda.}
\item{no_common}{Value indicating whether the component was a common component (1) or not (0).}
}
\description{
This function is used to bootstrap SNAC analysis and should not be used on its own.
}
