% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootSNAC_sum.R
\name{bootSNAC_sum}
\alias{bootSNAC_sum}
\title{Summarize additional bootstrapped SNAC network results}
\usage{
bootSNAC_sum(Bootstraps)
}
\arguments{
\item{Bootstraps}{A bootnet object containing the bootstrapped network estimation of a SNAC network}
}
\value{
\item{cocorec}{The common component recovery rate. How often is the component a common component?}
\item{lambdaCom}{The values for the lasso's lambda parameter (SNAC's component step)}
\item{lambdaNet}{The values for the graphical lasso's lambda parameter (SNAC's network step)}
\item{Structure}{A list of the component structures for each bootstrap result}
}
\description{
Summarize additional bootstrapped SNAC network results
}
