% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDpre.R
\name{CDpre}
\alias{CDpre}
\title{Variable selection algorithm with a predefined component loading structure}
\usage{
CDpre(DATA, Jk, R, Posit, GroupStructure, LASSO, MaxIter)
}
\arguments{
\item{DATA}{The concatenated data block, with rows representing subjects.}

\item{Jk}{A vector. Each element of this vector is the number of columns of a
data block.}

\item{R}{The number of components (R>=2).}

\item{Posit}{Indicate on which component(s) the Lasso Penalty is imposed. If unspecified, the algorithm assume that the
Lasso penalty is imposed on the common component(s) only. If there is no common component, then Lasso penalty is applied to all components.}

\item{GroupStructure}{XXX}

\item{LASSO}{A Lasso tuning parameter.}

\item{MaxIter}{Maximum number of iterations for this algorithm. The default
value is 400.}
}
\value{
Component loadings ($Pmatrix) and scores ($Tmatrix)
}
\description{
Note that parameter "Posit" indicate which components need a Lasso penalty. From R package \code{RegularizedSCA} (Gu & Van Deun, 2016).
}
\references{
Gu, Z., & Van Deun, K. (2016). A variable selection method for simultaneous component based data integration. \emph{Chemometrics and Intelligent Laboratory Systems}, 158, 187-199.
}
