% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkInput.R
\name{checkInput}
\alias{checkInput}
\title{Function from bootnet package (Epskamp, Borsboom, & Fried, 2018)}
\usage{
checkInput(
  default = c("none", "EBICglasso", "ggmModSelect", "pcor", "IsingFit", "IsingSampler",
    "huge", "adalasso", "mgm", "relimp", "cor", "TMFG", "ggmModSelect", "LoGo",
    "graphicalVAR", "piecewiseIsing", "SVAR_lavaan", "GGMncv"),
  fun,
  verbose = TRUE,
  .dots = list(),
  ...
)
}
\description{
Function from bootnet package (Epskamp, Borsboom, & Fried, 2018)
}
\references{
Epskamp, S., Borsboom, D., & Fried, E. I. (2018). Estimating psychological networks and their accuracy: A tutorial paper. Behavior Research Methods, 50(1), 195-212.
}
