% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comSNAC.R
\name{comSNAC}
\alias{comSNAC}
\title{Decompose multi-domain data into common and domain-specific components.}
\usage{
comSNAC(data, b1, b2, R, method = "structured", Target)
}
\arguments{
\item{data}{Data file consisting of 2 data blocks with subjects as rows and variables as columns. Variables should be grouped per data block}

\item{b1}{The number of variables in data block 1}

\item{b2}{The number of variables in data block 2}

\item{R}{The number of components}

\item{method}{How is component structure determined? Default is "structured" for which a predetermined component structure in the form of a Target is required.}

\item{Target}{Required for structured analysis. Should be matrix of 2 rows (for each data block) by R components. Enter 0 if data block should not contribute to that component; 1 is data block should contribute to that component.}
}
\value{
Component loadings ($Pmartix) and scores ($Tmatrix) of all R components. Additionally returns the tuning parameter lambda (estimated using 10-fold cross validation) and the predetermined component structure.
}
\description{
A component analysis technique that transforms a data set consisting of two data blocks (domains) into \code{R} components.
These components are either domain-specific (component loadings from only one domain) or common (component loadings from both domains); the latter of which can be used in \link{netSNAC}.
Part of Sparse Network And Component (SNAC) analysis. For more information on the component analysis see Gu & Van Deun (2016).
}
\references{
Gu, Z., & Van Deun, K. (2016). A variable selection method for simultaneous component based data integration. \emph{Chemometrics and Intelligent Laboratory Systems}, 158, 187-199.

Tio, P., Waldorp, L., & VanDeun, K. (2020). Constructing graphical models for multi-source data: Sparse network and component analysis. In \emph{Advanced Studies in Classification and Data Science} (pp. 275-287). Springer Singapore.
}
