% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_graphicalLasso.R
\name{cv_graphicalLasso}
\alias{cv_graphicalLasso}
\title{A k-fold cross-validation procedure for optimal graphical lasso tuning parameter}
\usage{
cv_graphicalLasso(
  data,
  nfolds = 10,
  GraphicalLasso = exp(seq(from = log(1e-07), to = log(0.05), length.out = 100))
)
}
\arguments{
\item{data}{A data matrix containing data of a single component (Tmatrix\%*\%t(Pmatrix))}

\item{nfolds}{Number of folds. If missing, 10-fold cross-validation will be performed}

\item{GraphicalLasso}{The range of graphical lasso tuning parameters. Default is a sequence of 100 numbers from 0.0000001 to 0.05.}
}
\value{
A graphical lasso tuning parameter that leads to a moddel with PRESS closest to the lowest PRESS + 1SE.
}
\description{
A k-fold cross-validation procedure for optimal graphical lasso tuning parameter
}
