% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determinant_check.R
\name{determinant_check}
\alias{determinant_check}
\title{Check whether a determinant of an inverse covariance matrix can be determined without warning or error}
\usage{
determinant_check(x)
}
\arguments{
\item{x}{Inverse covariance matrix}
}
\value{
TRUE if determinant can be determined without warnings or error
}
\description{
Check whether a determinant of an inverse covariance matrix can be determined without warning or error
}
