% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netSNAC.R
\name{netSNAC}
\alias{netSNAC}
\title{Estimate partial correlation network of common component}
\usage{
netSNAC(Pcommon, Tcommon, lasso = 0.002)
}
\arguments{
\item{Pcommon}{Component loadings of the common component(s). Should be a matrix of p variables by Rc number of common components}

\item{Tcommon}{Component scores of the common component(s). Should be a matrix of n participants by Rc number of common components}

\item{lasso}{Lambda parameter for graphical lasso. Default is set at 0.002}
}
\value{
An inverse covariance matrix and partial correlation matrix of common component
}
\description{
The partial correlation and inverse covariance matrices of the common component(s) are estimated using graphical lasso.
Part of Sparse Network And Component (SNAC) analysis, see also \link{comSNAC}.
}
\references{
Tio, P., Waldorp, L., & VanDeun, K. (2020). Constructing graphical models for multi-source data: Sparse network and component analysis. In \emph{Advanced Studies in Classification and Data Science} (pp. 275-287). Springer Singapore.
}
