% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NIMH_stand}
\alias{NIMH_stand}
\title{Standardized subset of National Institute of Mental Health's (NIMH) Volunteer Dataset}
\format{
A data frame with 183 rows and 31 variables:
\describe{
\item{Chloride}{A numeric vector representing chloride MMOL/L.}
\item{Creatine}{A numeric vector representing creatine MG/DL.}
\item{Glucose}{A numeric vector representing glucose MG/DL.}
\item{Potassium}{A numeric vector representing potassium MMOL/L}
\item{Sodium}{A numeric vector representing sodium MMOL/L}
\item{CO2}{A numeric vector representing total carbon dioxide MMOL/L}
\item{BUN}{A numeric vector representing blood urea nitrogen MG/DL}
\item{ALT}{A numeric vector representing alanine aminotransferase U/L}
\item{ALP}{A numeric vector representing alkaline phosphatase U/L}
\item{AST}{A numeric vector representing aspartate aminotransferase U/L}
\item{Bilirubin}{A numeric vector representing bilirubin total MG/DL}
\item{Basophil}{A numeric vector representing basophils \%}
\item{Eosinophil}{A numeric vector representing eosinophils \%}
\item{Hematocrit}{A numeric vector representing hematorcrit \%}
\item{Hb}{A numeric vector representing hemoglobin G/DL}
\item{Lymphocytes}{A numeric vector representing lymphocytes \%}
\item{MCH}{A numeric vector representing mean corpuscular hemoglobin PG}
\item{MCHC}{A numeric vector representing mean corpuscular hemoglobin concentration G/DL}
\item{MCV}{A numeric vector representing mean corpuscular volume FL}
\item{Monocytes}{A numeric vector representing monocytes \%}
\item{MPV}{A numeric vector representing mean platelet volume FL}
\item{Neutrophils}{A numeric vector representing neutrophils \%}
\item{RBC}{A numeric vector representing red blood count distribution width \%}
\item{CK}{A numeric vector representing creatine kinase U/L}
\item{CRP}{A numeric vector representing C-reactive protein, high sensitivity, comprehensive MG/L}
\item{TSH}{A numeric vector representing thyroid stimulating hormone MCIU/ML}
\item{Alcohol}{A numeric vector representing how often subject had a drink containing alcohol (0 - never; 1 - monthly or less; 2 - 2-4 times a week; 3 - 2-3 times a week; 4 - 4 or more times a week).}
\item{Verbal_IQ}{A numeric vector representing standardized verbal IQ score}
\item{Nonverbal_IQ}{A numeric vector representing standardized non-verbal IQ score}
\item{ME_fat}{A numeric vector representing mental fatigue score}
\item{ME_per}{A numeric vector representing mental performance}
}
}
\source{
\url{https://doi.org/doi:10.18112/openneuro.ds004215.v1.0.2}
}
\usage{
NIMH_stand
}
\description{
A subset of the National Institute of Mental Health's Healthy Volunteer data set (version 1.0.2) that contains a "biology" block (hepatic panel, chemistry panel, blood count, and blood chemistry measures) and a "psychology" block (AUDIT (alcohol use disorder), KBIT-2 (intelligence), and VAS (mental effort) measures).
All variables have been nonparanormal transformed (using hpn:huge.npn()).
All variables have been standardized and missing data has been handled using multiple imputations by chained equations (mice; Van Buuren & Groothuis-Oudshoorn, 2011) using SNACR::pre_process().
}
\references{
Nugent, A. C., Thomas, A. G., Mahoney, M., Gibbons, A., Smith, J. T., Charles, A. J., ... & Chung, J. (2022). The NIMH intramural healthy volunteer dataset: A comprehensive MEG, MRI, and behavioral resource. \emph{Scientific Data}, 9(1), 518.

Van Buuren, S., & Groothuis-Oudshoorn, K. (2011). mice: Multivariate imputation by chained equations in R. \emph{Journal of statistical software}, 45, 1-67.
}
\keyword{datasets}
