% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_common.R
\name{num_common}
\alias{num_common}
\title{Determine the component structure after the component step of Sparse Network And Component analysis}
\usage{
num_common(Pmatrix, Jk)
}
\arguments{
\item{Pmatrix}{P matrix of the component}

\item{Jk}{A vector with the number of columns of each data block}
}
\value{
Table with for each component whether they are a common, block 1-specific, or block 2-specific component.
}
\description{
Determine the component structure after the component step of Sparse Network And Component analysis
}
