% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statTable.R
\name{statTable}
\alias{statTable}
\title{Compute tidy table from bootnetResult object. Taken from bootnet package (Epskamp, Borsboom, & Fried, 2018)}
\usage{
statTable(
  x,
  name,
  alpha = 1,
  computeCentrality = TRUE,
  statistics = c("edge", "strength", "closeness", "betweenness"),
  directed = FALSE,
  bridgeArgs = list(),
  includeDiagonal = FALSE,
  ...
)
}
\arguments{
\item{x}{X}

\item{name}{Name}

\item{alpha}{Default is 1}

\item{computeCentrality}{Compute centrality? Default is TRUE}

\item{statistics}{Which statistics should be calculated}

\item{directed}{Is directed? Default is FALSE}

\item{bridgeArgs}{list}

\item{includeDiagonal}{Include diagonal? Default is FALSE}

\item{...}{arguments for function}
}
\value{
Result in data frame with entries: original (logical), name, type, node1, node2, value
}
\description{
Compute tidy table from bootnetResult object. Taken from bootnet package (Epskamp, Borsboom, & Fried, 2018)
}
\references{
Epskamp, S., Borsboom, D., & Fried, E. I. (2018). Estimating psychological networks and their accuracy: A tutorial paper. Behavior Research Methods, 50(1), 195-212.
}
