% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateNetwork_SNAC.R
\name{estimateNetwork_SNAC}
\alias{estimateNetwork_SNAC}
\title{This function takes data as input and produced a network. It is used inside bootnet_SNAC()}
\usage{
estimateNetwork_SNAC(
  data,
  default = c("none", "EBICglasso", "pcor", "IsingFit", "IsingSampler", "huge",
    "adalasso", "mgm", "relimp", "cor", "TMFG", "ggmModSelect", "LoGo", "graphicalVAR",
    "piecewiseIsing", "SVAR_lavaan", "GGMncv"),
  fun,
  labels,
  verbose = TRUE,
  .dots = list(),
  weighted = TRUE,
  signed = TRUE,
  directed,
  datatype,
  checkNumeric = FALSE,
  ...,
  .input,
  memorysaver = FALSE
)
}
\arguments{
\item{data}{Data matrix}
}
\description{
This function takes data as input and produced a network. It is used inside bootnet_SNAC()
}
