% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netSNAC.R
\name{netSNAC}
\alias{netSNAC}
\title{Estimate partial correlation network of common component}
\usage{
netSNAC(Pcommon, Tcommon, lambda_net)
}
\arguments{
\item{Pcommon}{Component loadings of the common component(s). Should be a matrix of p variables by Rc number of common components}

\item{Tcommon}{Component scores of the common component(s). Should be a matrix of n participants by Rc number of common components}

\item{lambda_net}{Range of the lambda tuning parameter for the graphical lasso. Default value is a sequence of 100 numbers from 0.0000001 to 0.05. The numbers are equally spaced on a log scale.}
}
\value{
The partial correlation matrix and inverse covariance matrix of the common component, as well as the
\item{graph}{Partial correlation matrix.}
\item{inv_cov}{Inverse covariance matrix.}
\item{lambda_net}{Lambda tuning parameter used in for graphical lasso that leads to the model with PRESS closest to the lowest PRESS + 1SE.}
}
\description{
The partial correlation and inverse covariance matrices of the common component(s) are estimated using graphical lasso.
Part of Sparse Network And Component (SNAC) analysis, see also \link{comSNAC}.
}
\references{
Tio, P., Waldorp, L., & VanDeun, K. (2020). Constructing graphical models for multi-source data: Sparse network and component analysis. In \emph{Advanced Studies in Classification and Data Science} (pp. 275-287). Springer Singapore.
}
